+++ tap_yotpo/streams/__init__.py
@@ -1,14 +1,14 @@
 """tap-yotpo streams module."""
+from .collections import Collections
+from .customers import Customers
 from .emails import Emails
+from .order_fulfillments import OrderFulfillments
 from .orders import Orders
 from .product_reviews import ProductReviews
+from .product_variants import ProductVariants
 from .products import Products
 from .reviews import Reviews
 from .unsubscribers import Unsubscribers
-from .collections import Collections
-from .customers import Customers
-from .product_variants import ProductVariants
-from .order_fulfillments import OrderFulfillments
 
 STREAMS = {
     Emails.tap_stream_id: Emails,
@@ -20,5 +20,5 @@ STREAMS = {
     Reviews.tap_stream_id: Reviews,
     Unsubscribers.tap_stream_id: Unsubscribers,
     Collections.tap_stream_id: Collections,
-    Customers.tap_stream_id: Customers
+    Customers.tap_stream_id: Customers,
 }