+++ tap_yotpo/streams/orders.py
@@ -5,12 +5,12 @@ from singer import Transformer, get_logger, metrics, write_record
 from singer.utils import strftime, strptime_to_utc
 
 from ..helpers import ApiSpec
-from .abstracts import IncremetalStream, UrlEndpointMixin
+from .abstracts import IncrementalStream, UrlEndpointMixin
 
 LOGGER = get_logger()
 
 
-class Orders(IncremetalStream, UrlEndpointMixin):
+class Orders(IncrementalStream, UrlEndpointMixin):
     """class for Orders stream."""
 
     stream = "orders"