+++ CHANGELOG.md
@@ -31,7 +31,7 @@
   * Update stream selection to use metadata rather than deprecated annotated-schema [#4](https://github.com/singer-io/tap-yotpo/pull/4)
 
 ## 1.0.1
-  * Update version of `requests` to `2.20.0` in response to CVE 2018-18074 
+  * Update version of `requests` to `2.20.0` in response to CVE 2018-18074
 
 ## 1.0.0
   * Version bump for initial release